new_backend/requirements.txt
fastapi==0.68.0

uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==4.0.39
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0.post1
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
requests==2.31.0
openai==1.55.3
beautifulsoup4==4.12.2
python-dateutil==2.8.2 
gunicorn==21.2.0
sendgrid==6.11.0 
# For scheduling follow-up and last-chance jobs
apscheduler==3.10.4

# For Gmail OAuth & sending
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0

